% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{AddDateTags}
\alias{AddDateTags}
\title{Adding date derived tags.}
\usage{
AddDateTags(data, dateColumnName)
}
\arguments{
\item{data}{a data frame}

\item{dateColumnName}{a date column over which the aggregation is done}
}
\description{
Aggregate values for given column names.
}
