% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{ColMaxes}
\alias{ColMaxes}
\title{Sparse matrix column maximums.}
\usage{
ColMaxes(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\description{
Find the column maximums in a sparse matrix.
}
\details{
The argument \code{smat} is expected to be of type "dgCMatrix".
}
