% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{ConvertMultiColumnDataFrameToSparseMatrix}
\alias{ConvertMultiColumnDataFrameToSparseMatrix}
\title{Convert multi-column data frame into a sparse matrix.}
\usage{
ConvertMultiColumnDataFrameToSparseMatrix(multiColDF, itemColName,
  tagTypeColNames)
}
\arguments{
\item{tagTypeColNames}{names of the tag type column names}

\item{Multi-column}{data frame id-tag relationship}

\item{idColName}{the column name of the item ID}
}
\description{
Convert multi-column data frame into a sparse matrix.
}
\details{
This does not work if the tagTypeColNames have dash in them.
I assume because of the string-to-formula conversion in SMRCreateItemTagMatrix.
Obviously, the dependence of the SMRCreateItemTagMatrix can be removed.
}
