% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIngestionFrameworkFunctions.R
\name{EmptyComputationSpecificationRow}
\alias{EmptyComputationSpecificationRow}
\title{Empty computation specification row.}
\usage{
EmptyComputationSpecificationRow()
}
\value{
A data frame with one row.
}
\description{
Gives a data frame with an "empty" computation specification row.
}
