% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonComputeFormula}
\alias{ERTMonComputeFormula}
\title{Compute specified formula.}
\usage{
ERTMonComputeFormula(ertObj, formulaSpec)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{formulaSpec}{A formula specification.}
}
\value{
An ERTMon object.
}
\description{
Computes a formula with a given specification using the feature 
sub-matrices of an ERTMon object.
}
\details{
The column names of \code{formulaSpec} are expected to include:
\code{c("TermID", "TermCoefficient", "FeatureName", "ReduceFunction", "Coefficient", "Exponent", "RatioPart")}.
The result matrix is assigned into \code{ertObj$Value}.
}
