% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMonCreateInterface.R
\name{ERTMonCreateInterface}
\alias{ERTMonCreateInterface}
\title{Event records transformation dashboard}
\usage{
ERTMonCreateInterface(
  dataDirectoryName = "./",
  testDataDirectoryName = "./",
  compSpecFileName = file.path(dataDirectoryName, "computationSpecification.csv"),
  exportModelID = as.character(Sys.time()),
  exportFilePrefix = gsub(pattern = " ", replacement = "_", paste0(exportModelID, "_"),
    fixed = TRUE)
)
}
\arguments{
\item{dataDirectoryName}{A directory name with data.}

\item{testDataDirectoryName}{A directory name with test data.}

\item{compSpecFileName}{A file name of computation specification; 
(a CSV file).}

\item{exportModelID}{A string with a model ID to be used for export.}

\item{exportFilePrefix}{A string with the prefix to be used for 
the export model files.}
}
\value{
Shiny app.
}
\description{
Creates the Shiny application with an 
event records transformations interface.
}
\seealso{
Other Interface functions: 
\code{\link{ERTMonMakeServerFunction}()},
\code{\link{ERTMonMakeUI}()}
}
\concept{Interface functions}
