% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonDataCheck}
\alias{ERTMonDataCheck}
\title{Check presence of required data.}
\usage{
ERTMonDataCheck(ertObj, functionName = NULL, logicalResult = FALSE)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{functionName}{A name of the delegating function (if any).}

\item{logicalResult}{Should the result be logical value?}
}
\value{
If \param(logicalValue) is FALSE the result is ERTMon object; if TRUE the result is logical value.
}
\description{
Checks does an ERTMon object have event records, entity attributes, and computation specification.
}
