% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonDataTransformerCheck}
\alias{ERTMonDataTransformerCheck}
\title{Check presence of a data transformer object.}
\usage{
ERTMonDataTransformerCheck(ertObj, functionName = "",
  logicalResult = FALSE)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{functionName}{The name of the delegating function.}

\item{logicalResult}{Should the result be logical value?}
}
\value{
A logical value or an ERTMon object.
}
\description{
Checks the presence of a data transformation S4 object in a given ERTMon object.
}
