% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonEmptyComputationSpecification}
\alias{ERTMonEmptyComputationSpecification}
\title{Empty computation specification.}
\usage{
ERTMonEmptyComputationSpecification(nrow = 1)
}
\value{
A data frame with one row.
}
\description{
Gives a data frame with an "empty" computation specification row.
}
\seealso{
Other Non-monadic functions.: 
\code{\link{ERTMonVerifyDataDirectory}()}
}
\concept{Non-monadic functions.}
