% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonEmptyComputationSpecificationRow}
\alias{ERTMonEmptyComputationSpecificationRow}
\title{Empty computation specification row.}
\usage{
ERTMonEmptyComputationSpecificationRow()
}
\value{
A data frame with one row.
}
\description{
Gives a data frame with an "empty" computation specification row.
}
\details{
Non-monadic functions.
}
