% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonExportToCSVFeatureMatrix}
\alias{ERTMonExportToCSVFeatureMatrix}
\title{Export the feature matrix into a CSV file.}
\usage{
ERTMonExportToCSVFeatureMatrix(ertObj, fileName = NULL, modelID = NULL)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{fileName}{A CSV file name. If \code{NULL} no file is written.}

\item{modelID}{A string; if \code{NULL} it is not used.}
}
\value{
An ERTMon object or \code{ERTMonFailureSymbol}.
}
\description{
Exports the feature matrix into a CSV file.
}
\details{
The value of the argument \code{modelID} (if not \code{NULL}) is
added as the first column to the feature matrix data frame (to be exported).
That data frame is assigned to the result monad object \code{$Value}.
}
