% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\docType{data}
\name{ERTMonFailureSymbol}
\alias{ERTMonFailureSymbol}
\title{Failure symbol for ERTMon.}
\format{An object of class \code{logical} of length 1.}
\usage{
ERTMonFailureSymbol
}
\description{
Failure symbol for the monad ERTMon.
}
\keyword{datasets}
