% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMonCreateInterface.R
\name{ERTMonMakeServerFunction}
\alias{ERTMonMakeServerFunction}
\title{Event records transformation server function}
\usage{
ERTMonMakeServerFunction()
}
\value{
Shiny server function
}
\description{
Creates the Shiny server function for an 
event records transformations interface
}
\seealso{
Other Interface functions: \code{\link{ERTMonCreateInterface}},
  \code{\link{ERTMonMakeUI}}
}
\concept{Interface functions}
