% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonProcessEventRecords}
\alias{ERTMonProcessEventRecords}
\title{Process the data and make feature matrix.}
\usage{
ERTMonProcessEventRecords(ertObj,
  outlierIdentifier = SPLUSQuartileIdentifierParameters,
  alignmentSpec = "MaxTime", echoStepsQ = FALSE,
  progressObject = NULL)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{outlierIdentifier}{Outlier parameters function.}

\item{alignmentSpec}{A time series alignment specification argument 
with acceptable values \code{"MinTime"}, \code{"MaxTime"}, or a non-negative number.}

\item{echoStepsQ}{Should the computational steps be proclaimed?}

\item{progressObject}{An object to be used in a progress gauge.}
}
\value{
An ERTMon object.
}
\description{
Processes the set event records using the set computation specification in an ERTMon object.
}
\details{
The result feature matrix is assigned to \code{ertObj$Value}.
If \code{alignmentSpec} is a non-negative number the alignment is done by finding 
the difference \code{ObservationTime - alignmentSpec} and keeping the event records for which
that difference is non-negative.
}
\seealso{
Other Transformation functions: \code{\link{ERTMonExtractFeatures}}
}
\concept{Transformation functions}
