% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonReadDataFromDirectory}
\alias{ERTMonReadDataFromDirectory}
\title{Read data from a directory.}
\usage{
ERTMonReadDataFromDirectory(ertObj, directoryName, readCompSpecQ = TRUE,
  progressObject = NULL, echoStepsQ = FALSE)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{directoryName}{A directory name.}

\item{progressObject}{An object to use for a progress gauge.}

\item{echoStepsQ}{Should the computational steps be proclaimed?}

\item{compSpec}{A computation specification data frame; if NULL it is read from the directory.}
}
\value{
An ERTMon object.
}
\description{
Reads event records, entity attributes, and computation specification
data frames from a directory.
}
\details{
The specified directory is expected to have the files 
\code{eventRecords.csv}, \code{entityAttributes.csv}, and \code{computationSpecification.csv}.
}
\seealso{
Other Data ingestion functions: \code{\link{ERTMonReadComputationSpecification}}
}
\concept{Data ingestion functions}
