% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonSetComputationSpecification}
\alias{ERTMonSetComputationSpecification}
\title{Set computation specification.}
\usage{
ERTMonSetComputationSpecification(ertObj, compSpec)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{compSpec}{A data frame that is a computation specification.}
}
\value{
An ERTMon object.
}
\description{
Sets a computation specification data frame into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeComputationSpecification}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeValue}}
}
\concept{Set/Take functions}
