% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonSetEntityAttributes}
\alias{ERTMonSetEntityAttributes}
\title{Set entity attributes.}
\usage{
ERTMonSetEntityAttributes(ertObj, eAttrs)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{eAttrs}{A data frame with entity attributes.}
}
\value{
An ERTMon object.
}
\description{
Sets an entity attributes data frame into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ERTMonSetComputationSpecification}()},
\code{\link{ERTMonSetEventRecords}()},
\code{\link{ERTMonTakeAlignmentSpec}()},
\code{\link{ERTMonTakeComputationSpecification}()},
\code{\link{ERTMonTakeContingencyMatrices}()},
\code{\link{ERTMonTakeEntityAttributes}()},
\code{\link{ERTMonTakeEventRecords}()},
\code{\link{ERTMonTakeFeatureMatrix}()},
\code{\link{ERTMonTakeFeatureNamePrefixes}()},
\code{\link{ERTMonTakeTimeCellsInterpretation}()},
\code{\link{ERTMonTakeTimeSeriesDataFrame}()},
\code{\link{ERTMonTakeTrasformedData}()},
\code{\link{ERTMonTakeValue}()},
\code{\link{ERTMonTakeVariableOutlierBoundaries}()}
}
\concept{Set/Take functions}
