% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonSetEventRecords}
\alias{ERTMonSetEventRecords}
\title{Set event records.}
\usage{
ERTMonSetEventRecords(ertObj, eRecs)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{eRecs}{A data frame with event records.}
}
\value{
An ERTMon object.
}
\description{
Sets an event records data frame into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonTakeComputationSpecification}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeTrasformedData}},
  \code{\link{ERTMonTakeValue}}
}
\concept{Set/Take functions}
