% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleTestDataCreation.R
\name{ERTMonSimpleTestData}
\alias{ERTMonSimpleTestData}
\title{Simple test data generation}
\usage{
ERTMonSimpleTestData(numberOfEntities = 10, numberOfVariables = 3,
  timeInterval = 900, numberOfTimeCells = 36,
  randomStartTimesQ = FALSE, variableFunction = "Linear",
  exportDirectoryName = NULL)
}
\arguments{
\item{numberOfEntities}{The number of entities.}

\item{numberOfVariables}{The number of variables}

\item{timeInterval}{The time interval corresponding to one time cell.}

\item{numberOfTimeCells}{The number of time cells.}

\item{randomStartTimesQ}{Should the starting times for different entities be different.}

\item{variableFunction}{A function or a function specification string, 
one of "Constant", "Linear", or "Random".}

\item{exportDirectoryName}{The name of the directory where the CSV files should be written.
If \code{NULL} the files are not exported}
}
\value{
A list with named elements.
}
\description{
Generates test data with specified parameters and writes the 
corresponding CSV files in a specified directory.
}
\details{
The element names corrspond to the exported files.
The element names of the result are "EntityAttributes", "EventRecords", "ComputationSpecification".
The function specified with \code{variableFunction} is applied to 
the points of a generated regular time grid.
}
\concept{Non-monadic functions}
