% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonStackFeatureMatrices}
\alias{ERTMonStackFeatureMatrices}
\title{Make a stack of feature sub-matrices.}
\usage{
ERTMonStackFeatureMatrices(
  ertObj,
  matrixNames = NULL,
  entityIDs = NULL,
  sep = ".",
  dropEmptyRowsQ = TRUE,
  completeColumnRangeQ = FALSE
)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{matrixNames}{A character vector with names of feature sub-matrices to be collapsed.
If NULL then all feature sub-matrices are collapsed.}

\item{entityIDs}{A character vector with entity ID's (that feature sub-matrices row names.)
If NULL then all entity ID's are used.}

\item{sep}{A separator string.}

\item{dropEmptyRowsQ}{Should the empty rows be dropped or not?}

\item{completeColumnRangeQ}{See \code{\link{ERTMonTakeContingencyMatrices}}.}
}
\description{
Make a stack of feature sub-matrices.
}
\details{
Stacks the feature sub-matrices (using \code{rbind}.)

The obtained matrix is assigned to \code{ertObj$Value}.
The entity ID's are concatenated with matrix names using the separator \code{sep}.
The rows that are "empty" (without elements) are removed by default.
}
\seealso{
Other Feature matrices: 
\code{\link{ERTMonCollapseFeatureMatrices}()},
\code{\link{ERTMonPlotFeatureMatrices}()}
}
\concept{Feature matrices}
