% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeAlignmentSpec}
\alias{ERTMonTakeAlignmentSpec}
\title{Take alignment specification.}
\usage{
ERTMonTakeAlignmentSpec(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame.
}
\description{
Returns the alignment specification used when 
the last time \code{ERTMonProcessEventRecords} was called.
}
\seealso{
Other Set/Take functions: 
\code{\link{ERTMonSetComputationSpecification}()},
\code{\link{ERTMonSetEntityAttributes}()},
\code{\link{ERTMonSetEventRecords}()},
\code{\link{ERTMonTakeComputationSpecification}()},
\code{\link{ERTMonTakeContingencyMatrices}()},
\code{\link{ERTMonTakeEntityAttributes}()},
\code{\link{ERTMonTakeEventRecords}()},
\code{\link{ERTMonTakeFeatureMatrix}()},
\code{\link{ERTMonTakeFeatureNamePrefixes}()},
\code{\link{ERTMonTakeTimeCellsInterpretation}()},
\code{\link{ERTMonTakeTimeSeriesDataFrame}()},
\code{\link{ERTMonTakeTrasformedData}()},
\code{\link{ERTMonTakeValue}()},
\code{\link{ERTMonTakeVariableOutlierBoundaries}()}
}
\concept{Set/Take functions}
