% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeComputationSpecification}
\alias{ERTMonTakeComputationSpecification}
\title{Take computation specification.}
\usage{
ERTMonTakeComputationSpecification(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame.
}
\description{
Takes the computation specification data frame from the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeValue}}
}
\concept{Set/Take functions}
