% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeEntityAttributes}
\alias{ERTMonTakeEntityAttributes}
\title{Take entity attributes.}
\usage{
ERTMonTakeEntityAttributes(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame.
}
\description{
Takes the entity attributes data frame from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ERTMonSetComputationSpecification}()},
\code{\link{ERTMonSetEntityAttributes}()},
\code{\link{ERTMonSetEventRecords}()},
\code{\link{ERTMonTakeAlignmentSpec}()},
\code{\link{ERTMonTakeComputationSpecification}()},
\code{\link{ERTMonTakeContingencyMatrices}()},
\code{\link{ERTMonTakeEventRecords}()},
\code{\link{ERTMonTakeFeatureMatrix}()},
\code{\link{ERTMonTakeFeatureNamePrefixes}()},
\code{\link{ERTMonTakeTimeCellsInterpretation}()},
\code{\link{ERTMonTakeTimeSeriesDataFrame}()},
\code{\link{ERTMonTakeTrasformedData}()},
\code{\link{ERTMonTakeValue}()},
\code{\link{ERTMonTakeVariableOutlierBoundaries}()}
}
\concept{Set/Take functions}
