% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeFeatureMatrix}
\alias{ERTMonTakeFeatureMatrix}
\title{Take the feature matrix.}
\usage{
ERTMonTakeFeatureMatrix(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A (sparse) matrix or \code{ERTMonFailureSymbol}.
}
\description{
Returns the feature matrix.
}
\details{
If the matrix does exists then \code{ERTMonFailureSymbol} is returned. 
The row names correspond to entities; the column names to features.
}
\seealso{
Other Set/Take functions: 
\code{\link{ERTMonSetComputationSpecification}()},
\code{\link{ERTMonSetEntityAttributes}()},
\code{\link{ERTMonSetEventRecords}()},
\code{\link{ERTMonTakeAlignmentSpec}()},
\code{\link{ERTMonTakeComputationSpecification}()},
\code{\link{ERTMonTakeContingencyMatrices}()},
\code{\link{ERTMonTakeEntityAttributes}()},
\code{\link{ERTMonTakeEventRecords}()},
\code{\link{ERTMonTakeFeatureNamePrefixes}()},
\code{\link{ERTMonTakeTimeCellsInterpretation}()},
\code{\link{ERTMonTakeTimeSeriesDataFrame}()},
\code{\link{ERTMonTakeTrasformedData}()},
\code{\link{ERTMonTakeValue}()},
\code{\link{ERTMonTakeVariableOutlierBoundaries}()}
}
\concept{Set/Take functions}
