% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeFeatureNamePrefixes}
\alias{ERTMonTakeFeatureNamePrefixes}
\title{Take feature name prefixes.}
\usage{
ERTMonTakeFeatureNamePrefixes(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A character vector.
}
\description{
Returns the prefixes of the column names of the feature matrix 
using the computation specification.
}
\details{
The matrix is not computed; only the computation specification is needed.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeValue}}
}
\concept{Set/Take functions}
