% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeTimeCellsInterpretation}
\alias{ERTMonTakeTimeCellsInterpretation}
\title{Take time cells interpretation.}
\usage{
ERTMonTakeTimeCellsInterpretation(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame.
}
\description{
Returns a data frame with the interpretation of the time cells.
}
\details{
If the alignment specification given to \code{\link{ERTMonProcessEventRecords}}
is \code{"MaxTime"} the entity time series will be aligned to finish at 0. 
In that case the interpretation times are non-positive.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeAlignmentSpec}},
  \code{\link{ERTMonTakeComputationSpecification}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeSeriesDataFrame}},
  \code{\link{ERTMonTakeTrasformedData}},
  \code{\link{ERTMonTakeValue}},
  \code{\link{ERTMonTakeVariableOutlierBoundaries}}
}
\concept{Set/Take functions}
