% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeTimeSeriesDataFrame}
\alias{ERTMonTakeTimeSeriesDataFrame}
\title{Take time series data frame.}
\usage{
ERTMonTakeTimeSeriesDataFrame(ertObj, origin = "1900-01-01")
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{origin}{A date-time object or something that can be coerced to such object,
that can be used in \code{\link{as.POSIXct}}.}
}
\value{
A data frame.
}
\description{
Returns a data frame with the feature matrix represented as
a set of time series in long form.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeAlignmentSpec}},
  \code{\link{ERTMonTakeComputationSpecification}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeTrasformedData}},
  \code{\link{ERTMonTakeValue}},
  \code{\link{ERTMonTakeVariableOutlierBoundaries}}
}
\concept{Set/Take functions}
