% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeTrasformedData}
\alias{ERTMonTakeTrasformedData}
\title{Take transformed data.}
\usage{
ERTMonTakeTrasformedData(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame or \code{ERTMonFailureSymbol}.
}
\description{
Takes the transformed data frame from the data transformer object 
which is in the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ERTMonSetComputationSpecification}()},
\code{\link{ERTMonSetEntityAttributes}()},
\code{\link{ERTMonSetEventRecords}()},
\code{\link{ERTMonTakeAlignmentSpec}()},
\code{\link{ERTMonTakeComputationSpecification}()},
\code{\link{ERTMonTakeContingencyMatrices}()},
\code{\link{ERTMonTakeEntityAttributes}()},
\code{\link{ERTMonTakeEventRecords}()},
\code{\link{ERTMonTakeFeatureMatrix}()},
\code{\link{ERTMonTakeFeatureNamePrefixes}()},
\code{\link{ERTMonTakeTimeCellsInterpretation}()},
\code{\link{ERTMonTakeTimeSeriesDataFrame}()},
\code{\link{ERTMonTakeValue}()},
\code{\link{ERTMonTakeVariableOutlierBoundaries}()}
}
\concept{Set/Take functions}
