% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeVariableOutlierBoundaries}
\alias{ERTMonTakeVariableOutlierBoundaries}
\title{Take variable outlier boundaries.}
\usage{
ERTMonTakeVariableOutlierBoundaries(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame.
}
\description{
Returns a data frame with the variable outlier boundaries.
}
\seealso{
Other Set/Take functions: 
\code{\link{ERTMonSetComputationSpecification}()},
\code{\link{ERTMonSetEntityAttributes}()},
\code{\link{ERTMonSetEventRecords}()},
\code{\link{ERTMonTakeAlignmentSpec}()},
\code{\link{ERTMonTakeComputationSpecification}()},
\code{\link{ERTMonTakeContingencyMatrices}()},
\code{\link{ERTMonTakeEntityAttributes}()},
\code{\link{ERTMonTakeEventRecords}()},
\code{\link{ERTMonTakeFeatureMatrix}()},
\code{\link{ERTMonTakeFeatureNamePrefixes}()},
\code{\link{ERTMonTakeTimeCellsInterpretation}()},
\code{\link{ERTMonTakeTimeSeriesDataFrame}()},
\code{\link{ERTMonTakeTrasformedData}()},
\code{\link{ERTMonTakeValue}()}
}
\concept{Set/Take functions}
