% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonUnit}
\alias{ERTMonUnit}
\title{Make a ERTMon Unit}
\usage{
ERTMonUnit(eventRecords = NULL, entityAttributes = NULL,
  compSpec = NULL)
}
\arguments{
\item{eventRecords}{A data frame with event records; can be NULL.}

\item{entityAttributes}{A data frame with entity attributes; can be NULL.}

\item{compSpec}{A data frame that is a computation specification; can be NULL.}
}
\value{
An S3 class "ERTMon". In other words, a list with the attribute "class" set to "ERTMon".
}
\description{
Creates a monad object.
}
