% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonVerifyDataDirectory}
\alias{ERTMonVerifyDataDirectory}
\title{Verify does a directory have ERTMon data files.}
\usage{
ERTMonVerifyDataDirectory(directoryName)
}
\arguments{
\item{directoryName}{A directory name string.}
}
\value{
A logical value.
}
\description{
Verify does a directory have the CSV files \code{eventRecords.csv} 
and \code{entityAttributes.csv}.
}
\seealso{
Other Non-monadic functions.: 
\code{\link{ERTMonEmptyComputationSpecification}()}
}
\concept{Non-monadic functions.}
