% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{FileColumnsIngest}
\alias{FileColumnsIngest}
\title{File columns ingestion.}
\usage{
FileColumnsIngest(fname, sep = "\\t", expectedColumns = 3,
  header = TRUE, apply.iconv = TRUE)
}
\arguments{
\item{fname}{filename assumed with suffix "TSV"}

\item{sep}{separator}

\item{header}{a logical value indicating whether the file contains the names of 
the variables as its first line}
}
\description{
Given a file name reads the data into column data frame
}
