% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{FileTriplets}
\alias{FileTriplets}
\title{File triplets.}
\usage{
FileTriplets(fname, sep = "\\t", propertiesToStrings = TRUE)
}
\arguments{
\item{fname}{file name}

\item{sep}{field separator}

\item{propertiesToStrings}{should all properties be turned into strings}
}
\description{
Turns the triplet records of a file into a sparse matrix
}
