% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{FileTripletsIngest}
\alias{FileTripletsIngest}
\title{File triplets ingestion.}
\usage{
FileTripletsIngest(fname, sep = "\\t")
}
\arguments{
\item{fname}{file name}

\item{sep}{separator}
}
\description{
Given a file name reads the data into three-column data frame, removes rows with NA and "null"
}
