% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{ImposeRowIDs}
\alias{ImposeRowIDs}
\title{Impose row ID's to a sparse matrix.}
\usage{
ImposeRowIDs(rowIDs, smat)
}
\arguments{
\item{rowIDs}{A character vector of row ID's.}

\item{smat}{A matrix with named rows.}
}
\value{
A sparse matrix.
}
\description{
Makes sure that the rows of a matrix are in 1-to-1 correspondence to an array of row ID's
}
