% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{IngestMultiValuedDataColumn}
\alias{IngestMultiValuedDataColumn}
\title{Multi-value column ingestion.}
\usage{
IngestMultiValuedDataColumn(itemRows, tagTypeColName,
  itemIDColName = "ID", nTagsPerField = 12, split = ",")
}
\arguments{
\item{itemRows}{a data frame of flat content data}

\item{tagTypeColName}{column name of the relationship to be ingested inge in itemRows}

\item{nTagsPerField}{number of tags per field of the column colName in itemRows}

\item{itemIDName}{}
}
\description{
Multi-value column ingestion.
}
