% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{MakeMatrixByColumnPartition}
\alias{MakeMatrixByColumnPartition}
\title{Make a matrix by column partitioning.}
\usage{
MakeMatrixByColumnPartition(data, colNameForRows, colNameForColumns,
  breaks = 10, leftOverlap = NULL, rightOverlap = NULL,
  colnamesPrefix = "")
}
\arguments{
\item{data}{a data frame}

\item{colNameForRows}{a column name in data for the rows of the result matrix}

\item{colNameForColumns}{a column name in data for the columns of the result matrix}

\item{breaks}{the points over which the breaking of data[colNameForColumns] is done}

\item{leftOverlap}{vector of weights for the neighboring columns to left}

\item{rightOverlap}{vector of weights for the neighboring columns to right}

\item{colnamesPrefix}{prefix for the columns names}
}
\description{
Make categorical representation of the numerical values of a column in a data frame and 
produce a matrix with the derived categorical tags as columns and values of a specified data column as rows.
}
