% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{RowMaxes}
\alias{RowMaxes}
\title{Sparse matrix row maximums.}
\usage{
RowMaxes(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\value{
A numeric vector with named elements
}
\description{
Find the row maximums in a sparse matrix.
}
\details{
The argument \code{smat} is expected to be of type "dgCMatrix".
This condition should hold \code{mean( names( RowMaxes(smat) ) == rownames(smat) ) == 1}.
}
