% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{SparseMatrixToTriplets}
\alias{SparseMatrixToTriplets}
\title{Convert sparse matrix into triplets}
\usage{
SparseMatrixToTriplets(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\value{
A data frame of triplets.
}
\description{
Converts a sparse matrix into a data frame triplets.
}
\seealso{
Other Sparse matrix transformation functions: \code{\link{ConvertMultiColumnDataFrameToSparseMatrix}},
  \code{\link{ToColumnValueIncidenceMatrix}},
  \code{\link{TripletsToSparseMatrix}}
}
\concept{Sparse matrix transformation functions}
