% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{SplitColumnOfTags}
\alias{SplitColumnOfTags}
\title{Split column of tags.}
\usage{
SplitColumnOfTags(dataColumn, numberOfSplits, sep = ",",
  emptyStringReplacement = "null")
}
\arguments{
\item{dataColumn}{data vector}

\item{numberOfSplits}{number of strings into which is field is splitted}

\item{emptyStringReplacement}{with what value empty strings are replaced}
}
\description{
Partition combined columns
}
