% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{SummaryPartitioned}
\alias{SummaryPartitioned}
\title{Summary by columns partitioning.}
\usage{
SummaryPartitioned(data, numberOfColumns = 3, ...)
}
\arguments{
\item{data}{a data frame}

\item{numberOfColumns}{number of columns for the partitioning}
}
\description{
Print the summary of data frame in a series of specified number of columns.
}
