% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{TagBasketsMatrixIntoItemTagMatrix}
\alias{TagBasketsMatrixIntoItemTagMatrix}
\title{Tag baskets matrix into item tag matrix}
\usage{
TagBasketsMatrixIntoItemTagMatrix(itemTagMat)
}
\arguments{
\item{itemTagMat}{item-tag table}
}
\description{
Tag table (an R-matrix) into binary incedence sparse matrix
}
