% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{TopOutlierPosition}
\alias{TopOutlierPosition}
\title{Top outlier positions finder.}
\usage{
TopOutlierPosition(dataVec,
  outlierIdentifier = HampelIdentifierParameters)
}
\arguments{
\item{dataVec}{data vector}

\item{outlierIdentifier}{outlier identifier function}
}
\description{
Find the top outlier positions in a data vector
}
