% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{TripletsToSparseMatrix}
\alias{TripletsToSparseMatrix}
\title{Make a sparse array/matrix from triplets.}
\usage{
TripletsToSparseMatrix(triplets)
}
\arguments{
\item{triplets}{a data frame with three columns}
}
\description{
Turns a data frame of three columns (triplets) into a sparse matrix
}
