% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{TwoFilesTriplets}
\alias{TwoFilesTriplets}
\title{Combine two triplets files.}
\usage{
TwoFilesTriplets(fname1, fname2, sep = "\\t",
  propertiesToStrings = TRUE)
}
\arguments{
\item{fname1}{name of the first file}

\item{fname2}{name of the second file}

\item{sep}{field separator}

\item{propertiesToStrings}{should the properties be converted into stings}
}
\description{
Combining the triplets of two files into a sparse matrix
}
