% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonComputeFormula}
\alias{ERTMonComputeFormula}
\title{Compute specified formula.}
\usage{
ERTMonComputeFormula(ertObj, formulaSpec, reduceFunc = "+")
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{formulaSpec}{A formula specification.}

\item{reduceFunc}{Reduction function, one of "+" or "*".}
}
\value{
An ERTMon object.
}
\description{
Computes a formula with a given specification using the feature 
sub-matrices of an ERTMon object.
}
\details{
The result matrix is assigned into ertObj$Value.
}
