% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonExportToCSVFeatureMatrix}
\alias{ERTMonExportToCSVFeatureMatrix}
\title{Export the feature matrix into a CSV file.}
\usage{
ERTMonExportToCSVFeatureMatrix(ertObj, fileName)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{fileName}{A CSV file name.}
}
\description{
Exports the feature matrix into a CSV file.
}
