% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonExtractFeatures}
\alias{ERTMonExtractFeatures}
\title{Extract feature with an alredy made data transformer.}
\usage{
ERTMonExtractFeatures(ertObj, eventRecords = NULL,
  entityAttribues = NULL, echoStepsQ = TRUE, progressObject = NULL)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{eventRecords}{A data frame with event records.}

\item{echoStepsQ}{Should the steps be echoed?}

\item{progressObject}{An object to be used in a progress guage.}

\item{entityAttributes}{A data frame with entity attributes.}
}
\value{
An ERTMon object.
}
\description{
Processes "unseen" data with an already made data transformer object.
}
\details{
The result feature matrix is assigned to \code{ertObj$Value}.
}
