% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonFilterEventRecords}
\alias{ERTMonFilterEventRecords}
\title{Filtering of event records.}
\usage{
ERTMonFilterEventRecords(ertObj, minObservationTime = NULL,
  maxObservationTime = NULL, locationIDs = NULL)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{locationIDs}{A character vector with location ID's.
If NULL no filtering over location ID's is done.}

\item{minObservatonTime}{Minimum observation time. 
It can be NULL, an integer alue (seconds), \code{"MinTime"}, or \code{"MaxMinTime"}.}

\item{maxObservatonTime}{Minimum observation time. 
NULL has the same effect as \code{"MinTime"} -- no filtering. 
It can be NULL, an integer alue (seconds), \code{"MaxTime"}, or \code{"MinMaxTime"}.
NULL has the same effect as \code{"MaxTime"} -- no filtering.}
}
\value{
An ERTMon object.
}
\description{
Filters event records for specified min and max observation times
and/or locations ID's.
}
\details{
The primary reason for this function is to assure the user
that all events adhere to certain explicitly postulated predicates.
\code{"MaxMinTime"} means that for each variable-and-entity pair is found the 
min observation time.
The found the min-max-times and max-min-times are assigned to 
\code{ertObj$SummarisedObservationTimes}.
}
