% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonPlotFeatureMatrices}
\alias{ERTMonPlotFeatureMatrices}
\title{Plot rows of feature sub-matrices as time series.}
\usage{
ERTMonPlotFeatureMatrices(ertObj, matrixNames = NULL, entityIDs = NULL,
  echoQ = TRUE, ...)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{matrixNames}{A character vector with matrix names to be plotted; NULL for all.}

\item{entityIDs}{A character vector with entity ID's to be plotted; NULL for all.≈}

\item{echoQ}{Should the result be plotted?}

\item{...}{Additional arguments for facet_wrap.}
}
\value{
An ERTMon object.
}
\description{
Plots the rows of the feature sub-matrices of an ERTMon object as time series.
}
