% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonProcessEventRecords}
\alias{ERTMonProcessEventRecords}
\title{Process the data and make feature matrix.}
\usage{
ERTMonProcessEventRecords(ertObj, echoStepsQ = TRUE,
  outlierIdentifier = SPLUSQuartileIdentifierParameters)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{echoStepsQ}{Should the steps be echoed?}

\item{outlierIdentifier}{Outlier parameters function.}
}
\value{
An ERTMon object.
}
\description{
Processes the set event records using the set computation specification in an ERTMon object.
}
