% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonSetComputationSpecification}
\alias{ERTMonSetComputationSpecification}
\title{Set computation specification.}
\usage{
ERTMonSetComputationSpecification(ertObj, compSpec)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{compSpec}{A data frame that is a computation specification.}
}
\value{
An ERTMon object.
}
\description{
Set a computation specification data frame into the monad object.
}
