% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonSetEntityAttributes}
\alias{ERTMonSetEntityAttributes}
\title{Set entity attributes.}
\usage{
ERTMonSetEntityAttributes(ertObj, eAttrs)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{eAttrs}{A data frame with entity attributes.}
}
\value{
An ERTMon object.
}
\description{
Set an entity attributes data frame into the monad object.
}
