% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonSetEventRecords}
\alias{ERTMonSetEventRecords}
\title{Set event records.}
\usage{
ERTMonSetEventRecords(ertObj, eRecs)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{eRecs}{A data frame with event records.}
}
\value{
An ERTMon object.
}
\description{
Set an event records data frame into the monad object.
}
