% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeContingencyMatrices}
\alias{ERTMonTakeContingencyMatrices}
\title{Take feature sub-matrices.}
\usage{
ERTMonTakeContingencyMatrices(ertObj, smat = NULL,
  columnPrefixesQ = FALSE)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{smat}{If not NULL it is going to be used instead of \code{ertObj$Value}.}

\item{columnPrefixesQ}{Should the column names have feature prefixes?
If TRUE the column names are just integers.}

\item{completeColumnRangeQ}{Should the columns correspond to all 
time grid cell indexes between the minimum and maximum one?
This is done only if \code{columnPrefixesQ=FALSE}.}
}
\value{
A list of named matrices.
}
\description{
Returns the sub-matrices of the feature matrix.
}
\details{
The sub-matrices are extracted through the corresponding prefixes.
An alternative is to use the sparse matrices that are in \code{ertObj$dtObj}.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeComputationSpecification}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeTrasformedData}},
  \code{\link{ERTMonTakeValue}},
  \code{\link{ERTMonTakeVariableOutlierBoundaries}}
}
\concept{Set/Take functions}
