% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeContingencyMatrices}
\alias{ERTMonTakeContingencyMatrices}
\title{Take feature sub-matrices.}
\usage{
ERTMonTakeContingencyMatrices(ertObj, smat = NULL,
  noColumnPrefixes = TRUE)
}
\arguments{
\item{ertObj}{An ERTMon object.}

\item{smat}{If not NULL it is going to be used instead of ertObj$Value.}

\item{noColumnPrefixes}{If TRUE the column names are just integers.}
}
\value{
A list of named matrices.
}
\description{
Returns the sub-matrices of the feature matrix.
}
\details{
The sub-matrices are extracted through the corresponding prefixes.
An alternative is to use the sparse matrices that are in ertObj$dtObj.
}
