% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeFeatureMatrix}
\alias{ERTMonTakeFeatureMatrix}
\title{Take the feature matrix.}
\usage{
ERTMonTakeFeatureMatrix(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A (sparse) matrix or ERTMonFailureSymbol.
}
\description{
Returns the feature matrix.
}
\details{
If the matrix does exists then ERTMonFailureSymbol is returned. 
The row names correspond to entities; the column names to features.
}
