% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeFeatureNamePrefixes}
\alias{ERTMonTakeFeatureNamePrefixes}
\title{Take feature name prefixes.}
\usage{
ERTMonTakeFeatureNamePrefixes(ertObj)
}
\arguments{
\item{ertObj}{an ERTMon object.}
}
\value{
A character vector.
}
\description{
Returns the prefixes of the column names of the feature matrix 
using the computation spefication.
}
\details{
The matrix is not computed; only the computation specification is needed.
}
