% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeTimeCellsInterpretation}
\alias{ERTMonTakeTimeCellsInterpretation}
\title{Take time cells interpretation.}
\usage{
ERTMonTakeTimeCellsInterpretation(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
A data frame.
}
\description{
Returns a data frame with the interpretation of the time cells.
}
\details{
Currently the entity time series can only be alligned to finish at 0. 
Hence the interpretation times are non-positive.
}
