% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeValue}
\alias{ERTMonTakeValue}
\title{Take the value in a ERTMon object.}
\usage{
ERTMonTakeValue(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
Just \code{ertObj$Value}.
}
\description{
Takes the value from ERTMon monad object.
}
\seealso{
Other Set/Take functions: \code{\link{ERTMonSetComputationSpecification}},
  \code{\link{ERTMonSetEntityAttributes}},
  \code{\link{ERTMonSetEventRecords}},
  \code{\link{ERTMonTakeComputationSpecification}},
  \code{\link{ERTMonTakeContingencyMatrices}},
  \code{\link{ERTMonTakeEntityAttributes}},
  \code{\link{ERTMonTakeEventRecords}},
  \code{\link{ERTMonTakeFeatureMatrix}},
  \code{\link{ERTMonTakeFeatureNamePrefixes}},
  \code{\link{ERTMonTakeTimeCellsInterpretation}},
  \code{\link{ERTMonTakeTrasformedData}}
}
\concept{Set/Take functions}
