% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERTMon.R
\name{ERTMonTakeValue}
\alias{ERTMonTakeValue}
\title{Take the value in a ERTMon object.}
\usage{
ERTMonTakeValue(ertObj)
}
\arguments{
\item{ertObj}{An ERTMon object.}
}
\value{
Just ertObj$Value.
}
\description{
Takes the value from ERTMon monad object.
}
