% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{SparseMatrixToTriplets}
\alias{SparseMatrixToTriplets}
\title{Convert a sparse matrix to triplets data frame.}
\usage{
SparseMatrixToTriplets(smat)
}
\arguments{
\item{smat}{a sparse matrix}
}
\value{
a data frame of triplets
}
\description{
Converts a sparse matrix to triplets
}
