% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunctions.R
\name{ToColumnValueIncidenceMatrix}
\alias{ToColumnValueIncidenceMatrix}
\title{Convert to incidence matrix by column values.}
\usage{
ToColumnValueIncidenceMatrix(mat, rowNames = TRUE, colNames = TRUE)
}
\arguments{
\item{mat}{an integer matrix to be converted to column value incidence matrix.}

\item{rowNames}{boolean to assign or not the result matrix row names to be the argument matrix row names}

\item{colNames}{boolean to assign or not the result matrix column names derived from the argument matrix column names}
}
\description{
Replaces each a column of a integer matrix with number of columns corresponding to the integer values.
The matrix [[2,3],[1,2]] is converted to [[0,0,1,0,0,0,0,1],[0,1,0,0,0,0,1,0]] .
}
